fix(js): renamed shared to indy-vdr-shared
Work funded by the Government of Ontario.
