+++ wrappers/javascript/indy-vdr-nodejs/package.json
@@ -29,7 +29,6 @@
     "build": "yarn clean && yarn compile",
     "clean": "rimraf -rf ./lib",
     "compile": "tsc -p tsconfig.build.json",
-    "example": "yarn --cwd example",
     "test": "jest",
     "install": "node-pre-gyp install"
   },
