+++ wrappers/javascript/indy-vdr-react-native/babel.config.js
@@ -1,7 +1,6 @@
-/* eslint-disable */
 const path = require('path')
 
-const indyVdrShared = require('../shared/package.json')
+const indyVdrShared = require('../indy-vdr-shared/package.json')
 
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
@@ -11,7 +10,7 @@ module.exports = {
       {
         extensions: ['.tsx', '.ts', '.js', '.json'],
         alias: {
-          [indyVdrShared.name]: path.join(__dirname, '../shared', indyVdrShared.source),
+          [indyVdrShared.name]: path.join(__dirname, '../indy-vdr-shared', indyVdrShared.source),
         },
       },
     ],
