+++ wrappers/javascript/indy-vdr-react-native/package.json
@@ -7,12 +7,12 @@
   "react-native": "build/index",
   "source": "src/index",
   "main": "build/index",
-  "homepage": "https://github.com/hyperledger/indy-vdr/tree/main/wrappers/javascript/react-native",
+  "homepage": "https://github.com/hyperledger/indy-vdr/tree/main/wrappers/javascript/indy-vdr-react-native",
   "author": "Hyperledger (https://github.com/hyperledger)",
   "repository": {
     "type": "git",
     "url": "https://github.com/hyperledger/indy-vdr",
-    "directory": "wrappers/javascript/react-native"
+    "directory": "wrappers/javascript/indy-vdr-react-native"
   },
   "publishConfig": {
     "access": "public"
