+++ wrappers/javascript/indy-vdr-shared/README.md
