+++ wrappers/javascript/indy-vdr-shared/src/builder/AcceptanceMechanismsRequest.ts
