+++ wrappers/javascript/indy-vdr-shared/src/builder/CredentialDefinitionRequest.ts
