+++ wrappers/javascript/indy-vdr-shared/src/builder/DisableAllTransactionAuthorAgreementsRequest.ts
