+++ wrappers/javascript/indy-vdr-shared/src/builder/GetAcceptanceMechanismsRequest.ts
