+++ wrappers/javascript/indy-vdr-shared/src/builder/GetAttribRequest.ts
