+++ wrappers/javascript/indy-vdr-shared/src/builder/GetCredentialDefinitionRequest.ts
