+++ wrappers/javascript/indy-vdr-shared/src/builder/GetNymRequest.ts
