+++ wrappers/javascript/indy-vdr-shared/src/builder/GetRevocationRegistryDefinitionRequest.ts
