+++ wrappers/javascript/indy-vdr-shared/src/builder/GetRevocationRegistryDeltaRequest.ts
