+++ wrappers/javascript/indy-vdr-shared/src/builder/GetRevocationRegistryRequest.ts
