+++ wrappers/javascript/indy-vdr-shared/src/builder/GetTransactionAuthorAgreementRequest.ts
