+++ wrappers/javascript/indy-vdr-shared/src/builder/GetTransactionRequest.ts
