+++ wrappers/javascript/indy-vdr-shared/src/builder/GetValidatorInfoAction.ts
