+++ wrappers/javascript/indy-vdr-shared/src/builder/RevocationRegistryDefinitionRequest.ts
