+++ wrappers/javascript/indy-vdr-shared/src/builder/RevocationRegistryEntryRequest.ts
