+++ wrappers/javascript/indy-vdr-shared/src/builder/TransactionAuthorAgreementRequest.ts
