+++ wrappers/javascript/indy-vdr-shared/src/builder/index.ts
