+++ wrappers/javascript/indy-vdr-shared/src/error.ts
