+++ wrappers/javascript/indy-vdr-shared/src/index.ts
