+++ wrappers/javascript/indy-vdr-shared/src/indyVdr/IndyVdrNativeBindings.ts
