+++ wrappers/javascript/indy-vdr-shared/src/indyVdr/index.ts
