+++ wrappers/javascript/indy-vdr-shared/src/indyVdr/indyVdr.ts
