+++ wrappers/javascript/indy-vdr-shared/src/types/IndyVdr.ts
