+++ wrappers/javascript/indy-vdr-shared/src/types/Utility.ts
