+++ wrappers/javascript/indy-vdr-shared/src/types/index.ts
