+++ wrappers/javascript/indy-vdr-shared/tsconfig.build.json
