+++ wrappers/javascript/indy-vdr-shared/tsconfig.json
