+++ wrappers/javascript/tsconfig.build.json
@@ -9,7 +9,7 @@
     "lib": ["ESNext"],
     "baseUrl": ".",
     "paths": {
-      "indy-vdr-shared": ["shared/src"],
+      "indy-vdr-shared": ["indy-vdr-shared/src"],
       "indy-vdr-react-native": ["indy-vdr-react-native/src"],
       "indy-vdr-nodejs": ["indy-vdr-nodejs/src"]
     },
